+++ src/lib/Sympa/List.pm
@@ -1494,17 +1494,12 @@ sub send_probe_to_user {
     my $who  = shift;
 
     # Shelve VERP for welcome or remind message if necessary
-    my $tracking;
-    if (    $self->{'admin'}{'welcome_return_path'} eq 'unique'
-        and $type eq 'welcome') {
-        $tracking = 'w';
-    } elsif ($self->{'admin'}{'remind_return_path'} eq 'unique'
-        and $type eq 'remind') {
-        $tracking = 'r';
-    } else {
-        #FIXME? Return-Path for '*_return_path' parameter with 'owner'
-        # value is LIST-owner address.  It might be LIST-request address.
-    }
+    my $tracking =
+        ($self->{'admin'}{'verp_welcome'} eq 'on' and $type eq 'welcome')
+        ? 'w'
+        : ($self->{'admin'}{'verp_remind'} eq 'on' and $type eq 'remind')
+        ? 'r'
+        : undef;
 
     my $spindle = Sympa::Spindle::ProcessTemplate->new(
         context  => $self,
@@ -5881,7 +5876,7 @@ sub add_list_header {
         }
     } elsif ($field eq 'archived_at') {
         if ($wwsympa_url and $self->is_web_archived()) {
-            # Use possiblly anonymized Message-Id: field instead of
+            # Use possibly anonymized Message-Id: field instead of
             # {message_id} attribute.
             my $message_id = Sympa::Tools::Text::canonic_message_id(
                 $message->get_header('Message-Id'));
@@ -5898,8 +5893,11 @@ sub add_list_header {
                 sprintf(
                     '<%s>',
                     Sympa::get_url(
-                        $self, 'arcsearch_id',
-                        paths => [$arc, $message_id]
+                        $self, 'msg',
+                        paths => [
+                            $arc,
+                            Sympa::Tools::Text::permalink_id($message_id)
+                        ]
                     )
                 )
             );
@@ -6299,7 +6297,7 @@ Limit result to the user with their e-mail $email.
 
 Returns:
 
-In array context, returns (possiblly empty or single-item) array of users.
+In array context, returns (possibly empty or single-item) array of users.
 In scalar context, returns reference to it.
 In case of database error, returns empty array or undefined value.
 
@@ -6424,7 +6422,7 @@ TBD.
 
 Returns:
 
-In array context, returns (possiblly empty or single-item) array of users.
+In array context, returns (possibly empty or single-item) array of users.
 In scalar context, returns reference to it.
 In case of database error, returns empty array or undefined value.