+++ src/lib/Sympa/Request/Handler/global_signoff.pm
@@ -50,7 +50,6 @@ sub _twist {
     my $auth_method =
           $request->{smime_signed} ? 'smime'
         : $request->{md5_check}    ? 'md5'
-        : $request->{dkim_pass}    ? 'dkim'
         :                            'smtp';
 
     my @target_lists;