+++ src/lib/Sympa/Spindle/AuthorizeRequest.pm
@@ -81,7 +81,6 @@ sub _twist {
     my $auth_method =
           $request->{smime_signed} ? 'smime'
         : $request->{md5_check}    ? 'md5'
-        : $request->{dkim_pass}    ? 'dkim'
         :                            'smtp';
 
     $result =