+++ src/lib/Sympa/Spindle/ToEditor.pm
@@ -89,19 +89,9 @@ sub _twist {
     );
 
     # Do not report to the sender if the message was tagged as a spam.
-    unless ($self->{quiet} or $message->{'spam_status'} eq 'spam') {
-        # Ensure 1 second elapsed since last message.
-        Sympa::send_file(
-            $list,
-            'message_report',
-            $sender,
-            {   type           => 'success',              # Compat. <=6.2.12.
-                entry          => 'moderating_message',
-                auto_submitted => 'auto-replied'
-            },
-            date => time + 1
-        );
-    }
+    Sympa::send_dsn($list, $message, {}, '2.3.0')
+        unless $self->{quiet}
+        or $message->{'spam_status'} eq 'spam';
     return 1;
 }