+++ src/lib/Sympa/WWW/Auth.pm
@@ -163,7 +163,7 @@ sub authentication {
             # Old style RC4 encrypted password.
             if ($user->{'password'} and $user->{'password'} =~ /\Acrypt[.]/) {
                 $log->syslog('notice',
-                    'Password in database seems encrypted. Run upgrade_sympa_password.pl to rehash passwords'
+                    'Password in database seems encrypted. Run \'sympa upgrade password\' to rehash passwords'
                 );
                 Sympa::send_notify_to_listmaster('*', 'password_encrypted',
                     {});